Update README to reflect changes
Last example still used "container.addDefinitions(...)" which is no longer part of the DIContainer class in the new v2. Also the IDIContainer interface was imported but not used.

Suggestion: Add a little notification at the top of the readme to let users know what changed between the new v2 and prior releases.